CDKによるデプロイ対応

+++ src/src/reply/index.ts
@@ -1,6 +1,6 @@
 import { GenericMessageEvent, SayFn } from '@slack/bolt'
 import { AwsCallback, AwsEvent } from '@slack/bolt/dist/receivers/AwsLambdaReceiver'
-import { createTables, runQuery } from '../common/dynamodb'
+import { runQuery } from '../common/dynamodb'
 import { awsLambdaReceiver, slackApp } from '../common/slack'
 import { RegisteredChannel } from './types'
 
@@ -35,19 +35,43 @@ async function getChannelData (
   }
 }
 
+async function isAlreadyReceived (message: GenericMessageEvent, say: SayFn): Promise<boolean> {
+  const messageChannel = message.channel
+  const messageTs = message.ts
+  const receivedMessages = await runQuery(
+    'SELECT channel, ts FROM youtube_streaming_watcher_received_messages WHERE channel=? AND ts=?',
+    [{ S: messageChannel }, { N: messageTs }]
+  )
+
+  if (receivedMessages?.Items !== undefined && receivedMessages.Items.length > 0) {
+    console.log(`message is already received: channel ${messageChannel}, ts ${messageTs}`)
+    return true
+  }
+
+  await say('処理中です。少々お待ちください :eyes:')
+  await runQuery(
+    'INSERT INTO youtube_streaming_watcher_received_messages VALUE {\'channel\': ?, \'ts\': ?}',
+    [{ S: messageChannel }, { N: messageTs }]
+  )
+  return false
+}
+
 export function setMessageEvents () {
-  slackApp.message('add', async ({
-    message,
-    say
-  }): Promise<void> => {
+  slackApp.message('add', async ({ message, say }): Promise<void> => {
+    message = message as GenericMessageEvent
+
+    if (await isAlreadyReceived(message, say)) {
+      return
+    }
+
     const channel = await getChannelData(message as GenericMessageEvent, say)
 
     if (channel === undefined) {
       return
     } else if (channel.exist) {
       await postMessage(
-        `このチャンネルは既に通知対象に追加されています: https://www.youtube.com/channel/${channel.id}`,
-        say
+                `このチャンネルは既に通知対象に追加されています: https://www.youtube.com/channel/${channel.id}`,
+                say
       )
       return
     }
@@ -57,37 +81,39 @@ export function setMessageEvents () {
       [{ S: channel.id }, { S: (new Date()).toISOString() }]
     )
     await postMessage(
-      `このチャンネルを通知対象に追加しました: https://www.youtube.com/channel/${channel.id}`,
-      say
+            `このチャンネルを通知対象に追加しました: https://www.youtube.com/channel/${channel.id}`,
+            say
     )
   })
 
-  slackApp.message('delete', async ({
-    message,
-    say
-  }): Promise<void> => {
-    const channel = await getChannelData(message as GenericMessageEvent, say)
+  slackApp.message('delete', async ({ message, say }): Promise<void> => {
+    message = message as GenericMessageEvent
+
+    if (await isAlreadyReceived(message, say)) {
+      return
+    }
+
+    const channel = await getChannelData(message, say)
 
     if (channel === undefined) {
       return
     } else if (!channel.exist) {
       await postMessage(
-        `このチャンネルは通知対象ではありません: https://www.youtube.com/channel/${channel.id}`,
-        say
+                `このチャンネルは通知対象ではありません: https://www.youtube.com/channel/${channel.id}`,
+                say
       )
       return
     }
 
     await runQuery('DELETE FROM youtube_streaming_watcher_channels WHERE channel_id=?', [{ S: channel.id }])
     await postMessage(
-      `このチャンネルを通知対象から削除しました: https://www.youtube.com/channel/${channel.id}`,
-      say
+            `このチャンネルを通知対象から削除しました: https://www.youtube.com/channel/${channel.id}`,
+            say
     )
   })
 }
 
 async function handler (event: AwsEvent, context: any, callback: AwsCallback) {
-  await createTables()
   setMessageEvents()
   const handler = await awsLambdaReceiver.start()
   return handler(event, context, callback)
